♻️ refactor: 알림 설정 로직 및 페이징 처리 구조 개선
- `design` 모듈에 `OnResumeEffect` Composable 유틸리티 추가
- `AlarmSettingScreen`의 라이프사이클 관찰 로직을 `OnResumeEffect`로 간소화
- `NotificationIntent` 정의를 별도 파일로 분리하여 모듈화
- `AlarmPagingSource`에 `safeApiCall`을 적용하여 에러 처리 및 응답 변환 로직 리팩토링
- Firebase 의존성을 Version Catalog Bundle로 통합하여 관리 효율성 증대
